+++ src/vue/fields/InputField.vue
@@ -103,12 +103,7 @@ export default {
         input: event => {
           this.normalizeTargetValue(event.target)
           if (this.value === event.target.value) return
-          if (this.type === INPUT_TYPES.number) {
-            // eslint-disable-next-line max-len
-            this.$emit(EVENTS.input, event.target.valueAsNumber.toString() || event.target.value)
-          } else {
-            this.$emit(EVENTS.input, event.target.value)
-          }
+          this.$emit(EVENTS.input, event.target.value)
         },
       }
     },