+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Please check our [developers guide](https://gitlab.com/tokend/developers-guide)
 for further information about branching and tagging conventions.
 
+
+## [Unreleased]
+#### Fixed
+- number input field
+
 ## [1.7.0-rc.3] - 2020-03-25
 #### Fixed
 - A bug with impossibility use a comma as a separator in Firefox brow